# src/main/resources/openapi/schema.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'VSA Projekt 2 schema'
  license:
    name: MIT
servers:
  - url: 'localhost:8080/api'
paths:
...
  /cars:
        get:
            summary: 'Return cars'
            operationId: getCars
            parameters:
                - name: user
                  required: false
                  in: query
                  schema:
                    type: integer
                    format: int64
                  description: 'Id of user - if given only cars belonging to that user will be returned'
                - name: vrp
                  in: query
                  required: false
                  description: 'Vehicle registration plate - if given, only vehicle that has this vrp will be returned'
                  schema:
                    type: string
            responses:
                '200':
                    description: 'Cars returned'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Car'
        post:
            summary: 'Create car'
            operationId: createCar
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Car' 
            responses:
                '201':
                    description: 'Car created and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Car'
                '400':
                    description: 'Incorrectly formed request body'
...
  /users:
        get:
            summary: 'Get users'
            operationId: getUsers
            parameters:
                - name: email
                  in: query
                  required: false
                  description: 'Email of user, if given only user with given email will be returned if it exists'
                  schema:
                      type: string
                      format: email
            responses:
                '200':
                    description: 'Users returned'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
        post:
            summary: 'Create user'
            operationId: createUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                '201':
                    description: 'User created and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Incorrectly formed request body'
  /users/{id}:
        get:
            summary: 'Get user with id'
            operationId: getUser
            parameters:
                - name: id
                  in: path
                  required: true
                  description: 'Id of user'
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: 'User exists and is returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '404':
                    description: 'User not found'
        put:
            summary: 'Update user'
            operationId: updateUser
            parameters:
                - name: id
                  in: path
                  required: true
                  description: 'Id of user'
                  schema:
                      type: integer
                      format: int64
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                           $ref: '#/components/schemas/User'
            responses:
                '200':
                    description: 'User updated and returned'
                    content:
                        application/json:
                            schema:
                               $ref: '#/components/schemas/User' 
                '404':
                    description: 'User not found'
                '400':
                    description: 'Incorrectly formed request body'
        delete:
            summary: 'Delete user'
            operationId: deletUser
            parameters:
                - name: id
                  in: path
                  required: true
                  description: 'Id of user'
                  schema:
                      type: integer
                      format: int64
            responses:
                '204':
                    description: 'User deleted'
                '404':
                    description: 'User does not exist'
...
  /cartypes:
        get:
            summary: 'Get car types'
            operationId: getCarTypes
            parameters:
                - name: name
                  required: false
                  in: query
                  description: 'Name of a cartype. If set only type with that name will be part of the collection'
                  schema:
                      type: string
            responses:
                '200':
                    description: 'Car types returned'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/CarType'
        
        post:
            summary: 'Create car type'
            operationId: createCarType
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                           $ref: '#/components/schemas/CarType'
            responses:
                '201':
                    description: 'Car type created and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarType'
                '400':
                    description: 'Incorrectly formed request body'
  /cartypes/{id}:
        get:
            summary: 'Get car type'
            operationId: getCarType
            parameters:
                - name: id
                  required: true
                  in: path
                  description: 'Id of the car type'
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: 'Car type found and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarType'
                '404':
                    description: 'Car type does not exist'
        delete:
            summary: 'Delete car type'
            operationId: deleteCarType
            parameters:
                - name: id
                  required: true
                  in: path
                  description: 'Id of the car type'
                  schema:
                      type: integer
                      format: int64
            responses:
                '204':
                    description: 'Car type deleted'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarType'
                '404':
                    description: 'Car type not found'
...
components:
  schemas:
...
    Car:
        type: object
        required:
            - id
            - brand
            - model
            - vrp
            - color
            - owner
        properties:
            id:
                type: integer
                format: int64
            brand:
                type: string
            model:
                type: string
            vrp:
                type: string
            colour:
                type: string
            type:
                type: object
                $ref: CarType
            owner:
                type: object
                $ref: UserDownFromCar
            reservations:
                type: array
                items:
                    type: object
                    $ref: ReservationDownFromCar
...
    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
            type: string
        email:
            type: string
        cars:
            type: array
            items:
                type: object
                $ref: CarDownFromUser
...
    CarType:
        type: object
        required:
            - id
            - name
        properties:
            id:
                type: integer
                format: int64
            name:
                type: string
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
security:
  - BasicAuth: []